+++ Client.Core/Internal/RestSharpExtensions.cs
@@ -10,9 +10,17 @@ namespace InfluxDB.Client.Core.Internal
 {
     internal static class RestSharpExtensions
     {
-        internal static IEnumerable<HeaderParameter> ToHeaderParameters(this HttpHeaders httpHeaders)
+        /// <summary>
+        /// Transform `HttpHeaders` to `HeaderParameter` type.
+        /// </summary>
+        /// <param name="httpHeaders"></param>
+        /// <param name="httpContentHeaders">Additionally content Headers</param>
+        /// <returns>IEnumerable&lt;HeaderParameter&gt;</returns>
+        internal static IEnumerable<HeaderParameter> ToHeaderParameters(this HttpHeaders httpHeaders,
+            HttpContentHeaders httpContentHeaders = null)
         {
             return httpHeaders
+                .Concat(httpContentHeaders ?? Enumerable.Empty<KeyValuePair<string, IEnumerable<string>>>())
                 .SelectMany(x => x.Value.Select(y => (x.Key, y)))
                 .Select(x => new HeaderParameter(x.Key, x.y));
         }