+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### Bug Fixes
 1. [#313](https://github.com/influxdata/influxdb-client-csharp/pull/313): Using default `org` and `bucket` in `WriteApiAsync`
+1. [#317](https://github.com/influxdata/influxdb-client-csharp/pull/317): Decompress Gzipped data
+1. [#317](https://github.com/influxdata/influxdb-client-csharp/pull/317): Logging HTTP headers from streaming request
 
 ### Documentation
 1. [#314](https://github.com/influxdata/influxdb-client-csharp/pull/314): Add Parameterized Queries example